fix: allChatRoom 파일 분리 및 warning 경고 수정
